refactor(docs): co-locate isActive utility and remove unused COMPONENTS.md
- Move is-active.ts to PageClient/utils (only used there)
- Delete apps/docs/COMPONENTS.md (no longer needed)
- Update import path in PageClient.tsx
