+++ apps/docs/src/app/(docs)/[[...slug]]/components/DocsPageLayout/components/PageClient/utils/is-active.ts
